pyproject.toml
[tool.poetry]
name = "api"
version = "0.3.0"
description = "Repository for Siege API Service."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
python-decouple = "^3.7"
itsdangerous = "^2.1.2"
django-guardian = "^2.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
commitizen = "^2.40.0"
flake8 = "^6.0.0"
coverage = "^7.1.0"
mypy = "^0.991"
django-stubs = {version = "^1.14.0", extras = ["compatible-mypy"]}
djangorestframework-stubs = {version = "^1.8.0", extras = ["compatible-mypy"]}

[tool.black]
color = true
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.commitizen]
version = "0.3.0"
version_files = ["pyproject.toml"]
update_changelog_on_bump = true
changelog_incremental = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
